requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
faker==24.11.0
requests==2.31.0
kafka-python==2.0.2
pyspark==3.5.1
clickhouse-connect
clickhouse-driver
fastapi
uvicorn
pydantic
pandas
APScheduler==3.10.4